create menu

+++ src/entites/Entity.java
@@ -1,9 +1,11 @@
 package entites;
 
-import java.awt.*;
+import java.awt.Color;
+import java.awt.Graphics;
 import java.awt.geom.Rectangle2D;
 
 public abstract class Entity {
+
     protected float x, y;
     protected int width, height;
     protected Rectangle2D.Float hitbox;
@@ -23,7 +25,7 @@ public abstract class Entity {
 
     }
 
-    protected void initHitbox(float x, float y, float width, float height) {
+    protected void initHitbox(float x, float y, int width, int height) {
         hitbox = new Rectangle2D.Float(x, y, width, height);
     }
